[C8] Split profile.go by responsibility (#2242)
model/sketch/profile.go was 614 LOC, over the 500-line ceiling. Split the
profile extraction from sketch geometry along its responsibilities; each file
is now under the ceiling:

- profile.go (256): the public profile model (ProfileEntity/Loop/Profile/
  Profiles), the Sketch.Profiles entry + geometry signature/normalisation and
  the closed-loop extraction.
- profile_trace.go (134): loop tracing — the directed edge set of a region
  boundary, the deterministic ring walk, and the coordinate welder.
- profile_nesting.go (163): containment/nesting classification — grouping the
  traced loops into profiles by containment, merging abutting loops, and the
  point-in-polygon / loop-containment predicates.
- profile_polygon.go (100): the circle-to-polygon conversion — sampling a
  sketch circle into a loop polygon (base N-gon, through-touch-points variant,
  angle de-duplication).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
per-file import redistribution differs); imports resolved per file with
goimports. Sketch package builds, archguard passes, lint 0. Net delta of
recognizers / tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2242.

(cherry picked from commit b0744dccb3f478b8ad58e34e97c24650bd6d6774)
