+++ model/sketch/profile.go
@@ -4,7 +4,6 @@ package sketch
 
 import (
 	stdmath "math"
-	"sort"
 
 	"oblikovati.org/math"
 )
@@ -254,361 +253,4 @@ func (s *Sketch) normalGeometry() []Entity {
 		out = append(out, e)
 	}
 	return out
-}
-
-// groupRegions builds profiles from closed loops by even–odd nesting: a loop contained in an even
-// number of others is an outer boundary; loops one level deeper inside it are its holes.
-//
-// As an exception, an odd-nested loop the user actually DREW also bounds its own selectable region
-// (the area inside it), so nested loops yield BOTH the annulus AND the inner face, matching how
-// Inventor lets you select either. A circle inside a rectangle offers the disk (#1526), and a
-// rounded-rectangle (or any polygon) inside another offers its inner face (#2165 — the offset of a
-// rounded profile left the inner region unselectable). Synthetic text-glyph counters carry NO
-// entities (they are bare polygons from TextProfiles), so they stay holes only and the glyph region
-// detection is unchanged; the grill boolean uses ClosedLoops (#863) and is likewise untouched.
-//
-// The extra inner regions are APPENDED after every outer region, so a loop the exception newly exposes
-// never shifts an outer region's index — index-based extrude/loft selections in older documents keep
-// pointing at the same face (newer ones resolve by seed point, #region-seed).
-func groupRegions(loops []Loop) []*Profile {
-	depth := make([]int, len(loops))
-	for i := range loops {
-		for j := range loops {
-			if i != j && containsLoop(loops[j], loops[i]) {
-				depth[i]++
-			}
-		}
-	}
-	var profiles []*Profile
-	for i, l := range loops {
-		if depth[i]%2 == 0 {
-			profiles = append(profiles, &Profile{outer: l, inner: holesOf(i, loops, depth)})
-		}
-	}
-	abutting := abuttingLoops(loops)
-	for i, l := range loops {
-		if depth[i]%2 == 1 && loopBoundsOwnRegion(l) && !abutting[i] {
-			profiles = append(profiles, &Profile{outer: l, inner: holesOf(i, loops, depth)})
-		}
-	}
-	return profiles
-}
-
-// loopBoundsOwnRegion reports whether an odd-nested (hole) loop ALSO bounds a face the user can
-// select in its own right. A loop the user drew — one that carries sketch entities, whether a single
-// circle/ellipse or a multi-entity polygon/rounded-rectangle chain — does; a synthetic text-glyph
-// counter (a bare polygon with no entities, produced by TextProfiles) does not, so the hole in an
-// 'O'/'A'/'B' stays a hole. See groupRegions (#1526, #2165).
-func loopBoundsOwnRegion(l Loop) bool {
-	return len(l.entities) >= 1
-}
-
-// abuttingLoops flags each loop that shares an edge with another loop — the minimal cells of one
-// SUBDIVIDED island (a grid of crossing grill bars carves an interior into many abutting cells). Such
-// a cell is NOT a region the user drew, so it stays a hole (merged into one outline by holesOf, #863);
-// only a STANDALONE inner loop — one sharing no edge with any sibling, i.e. a genuinely nested circle,
-// rectangle or rounded rectangle — bounds its own selectable region (#2165). Edges are welded so two
-// cells tracing the same boundary points register as shared.
-func abuttingLoops(loops []Loop) []bool {
-	edgeUses := map[[2]int]int{}
-	loopEdges := make([][][2]int, len(loops))
-	w := newLoopWelder()
-	for li, l := range loops {
-		loopEdges[li] = weldedLoopEdges(w, l)
-		for _, key := range loopEdges[li] {
-			edgeUses[key]++
-		}
-	}
-	abut := make([]bool, len(loops))
-	for li := range loops {
-		for _, key := range loopEdges[li] {
-			if edgeUses[key] > 1 {
-				abut[li] = true
-				break
-			}
-		}
-	}
-	return abut
-}
-
-// weldedLoopEdges returns a loop's undirected edges as sorted welded-index pairs, so an edge two
-// abutting cells trace from the same boundary points registers as the one shared edge.
-func weldedLoopEdges(w *loopWelder, l Loop) [][2]int {
-	idx := make([]int, len(l.polygon))
-	for i, p := range l.polygon {
-		idx[i] = w.add(p)
-	}
-	edges := make([][2]int, 0, len(idx))
-	for i := range idx {
-		a, b := idx[i], idx[(i+1)%len(idx)]
-		if a == b {
-			continue
-		}
-		if a > b {
-			a, b = b, a
-		}
-		edges = append(edges, [2]int{a, b})
-	}
-	return edges
-}
-
-// holesOf returns the holes of outer (index oi): the loops one nesting level inside it. A
-// disconnected island whose interior is itself subdivided (e.g. a grid of crossing bars inside
-// the boundary) arrives as many abutting minimal cells at that level; they are merged into the
-// island's outline loop(s) so the profile carries clean, non-abutting holes rather than a tiling
-// of overlapping cells (#863 — abutting hole loops are degenerate downstream).
-func holesOf(oi int, loops []Loop, depth []int) []Loop {
-	var holes []Loop
-	for j, l := range loops {
-		if j != oi && depth[j] == depth[oi]+1 && containsLoop(loops[oi], l) {
-			holes = append(holes, l)
-		}
-	}
-	return mergeAbuttingLoops(holes)
-}
-
-// mergeAbuttingLoops fuses loops that share edges (the minimal cells of one connected island)
-// into their boundary outline: a directed edge traversed by two abutting cells appears in both
-// directions and cancels; the surviving edges chain into the outline. Disjoint loops (separate
-// islands) share no edges and pass through unchanged.
-func mergeAbuttingLoops(loops []Loop) []Loop {
-	if len(loops) <= 1 {
-		return loops
-	}
-	w := newLoopWelder()
-	dir := map[[2]int]int{}
-	for _, l := range loops {
-		idx := make([]int, len(l.polygon))
-		for i, p := range l.polygon {
-			idx[i] = w.add(p)
-		}
-		for i := range idx {
-			if a, b := idx[i], idx[(i+1)%len(idx)]; a != b {
-				dir[[2]int{a, b}]++
-			}
-		}
-	}
-	rings := chainLoopBoundary(boundaryDirEdges(dir), w.points)
-	if len(rings) == 0 {
-		return loops // chaining failed (unexpected); keep the originals rather than drop holes
-	}
-	out := make([]Loop, len(rings))
-	for i, r := range rings {
-		out[i] = Loop{polygon: r, closed: true}
-	}
-	return out
-}
-
-// boundaryDirEdges keeps the directed edges whose reverse is absent — the outline of the union
-// (an edge shared by two abutting cells appears in both directions and is dropped).
-//
-// Emitted in sorted order: Go randomizes map iteration, and this slice seeds the adjacency
-// chainLoopBoundary walks, whose append order decides how edges pair at a touching vertex.
-func boundaryDirEdges(dir map[[2]int]int) [][2]int {
-	var keep [][2]int
-	for _, e := range sortedDirKeys(dir) {
-		if dir[[2]int{e[1], e[0]}] == 0 {
-			for k := 0; k < dir[e]; k++ {
-				keep = append(keep, e)
-			}
-		}
-	}
-	return keep
-}
-
-// sortedDirKeys returns the directed edges in ascending (from, to) order.
-func sortedDirKeys(dir map[[2]int]int) [][2]int {
-	keys := make([][2]int, 0, len(dir))
-	for e := range dir {
-		keys = append(keys, e)
-	}
-	sort.Slice(keys, func(i, j int) bool {
-		if keys[i][0] != keys[j][0] {
-			return keys[i][0] < keys[j][0]
-		}
-		return keys[i][1] < keys[j][1]
-	})
-	return keys
-}
-
-// chainLoopBoundary links directed edges head-to-tail into closed rings.
-//
-// The ring START is taken in sorted order, not map order. A ring's start vertex fixes the point
-// SEQUENCE Polygon() hands downstream, and the extrude builds one prism face per polygon edge in
-// that order — so a randomly-rotated hole ring produced a prism with the same vertices but its
-// faces on different planes, and the boolean consuming it drifted run to run (#23). Where two hole
-// rings touch at a vertex the start also decides the edge PAIRING, changing the rings themselves.
-func chainLoopBoundary(edges [][2]int, pts []math.Point2) [][]math.Point2 {
-	next := make(map[int][]int, len(edges))
-	for _, e := range edges {
-		next[e[0]] = append(next[e[0]], e[1])
-	}
-	starts := make([]int, 0, len(next))
-	for v := range next {
-		starts = append(starts, v)
-	}
-	sort.Ints(starts)
-	var rings [][]math.Point2
-	for _, start := range starts {
-		for len(next[start]) > 0 {
-			if ring := traceRing(start, next, pts); len(ring) >= 3 {
-				rings = append(rings, ring)
-			}
-		}
-	}
-	return rings
-}
-
-// traceRing walks next-edges from start until it returns, consuming each edge once.
-func traceRing(start int, next map[int][]int, pts []math.Point2) []math.Point2 {
-	ring := []math.Point2{pts[start]}
-	for cur := start; ; {
-		outs := next[cur]
-		if len(outs) == 0 {
-			return nil // open chain (degenerate)
-		}
-		nxt := outs[0]
-		next[cur] = outs[1:]
-		if nxt == start {
-			return ring
-		}
-		ring = append(ring, pts[nxt])
-		cur = nxt
-	}
-}
-
-// loopWelder merges coincident loop vertices onto a shared index list (a tolerance grid coarser
-// than the arrangement weld so computed cell corners coincide).
-type loopWelder struct {
-	index  map[[2]int64]int
-	points []math.Point2
-}
-
-func newLoopWelder() *loopWelder { return &loopWelder{index: map[[2]int64]int{}} }
-
-func (w *loopWelder) add(p math.Point2) int {
-	const grid = 1e-7 // tol:calibrated — 2D loop-welder coincidence grid (see brep arrange2d)
-	k := [2]int64{int64(stdmath.Round(p.X / grid)), int64(stdmath.Round(p.Y / grid))}
-	if i, ok := w.index[k]; ok {
-		return i
-	}
-	w.index[k] = len(w.points)
-	w.points = append(w.points, p)
-	return len(w.points) - 1
-}
-
-// containsLoop reports whether outer contains inner: every vertex of inner lies
-// inside outer's polygon. (Testing all vertices, not just the centroid, avoids a
-// false positive when a hole is centered on the region's centroid.)
-func containsLoop(outer, inner Loop) bool {
-	if len(inner.polygon) == 0 {
-		return false
-	}
-	for _, v := range inner.polygon {
-		if !pointInPolygon(v, outer.polygon) {
-			return false
-		}
-	}
-	return true
-}
-
-// pointInPolygon is the even–odd ray-casting test.
-func pointInPolygon(p math.Point2, poly []math.Point2) bool {
-	inside := false
-	n := len(poly)
-	for i, j := 0, n-1; i < n; j, i = i, i+1 {
-		yi, yj := poly[i].Y, poly[j].Y
-		if (yi > p.Y) != (yj > p.Y) {
-			x := poly[i].X + (p.Y-yi)/(yj-yi)*(poly[j].X-poly[i].X)
-			if p.X < x {
-				inside = !inside
-			}
-		}
-	}
-	return inside
-}
-
-// circleSamples is how many points a closed curve is sampled into for containment.
-const circleSamples = 24
-
-// sampleCircle returns a polygon approximating a circle.
-func sampleCircle(c *Circle) []math.Point2 {
-	return sampleCircleN(c, circleSamples)
-}
-
-// onCircleTol is how far a point may sit from a circle and still count as touching it. The touch is
-// exact in the authored geometry (a line constrained to end on the circle), so this only absorbs
-// float noise in re-deriving the radius; it matches the arrangement's own node merge.
-const onCircleTol = arrMergeTol // tol:calibrated — point-on-circle touch; see arrMergeTol
-
-// sampleCircleThrough samples a circle at the uniform angles PLUS the angle of every point that
-// lies ON it, so a curve ending there shares a vertex with the circle instead of missing it by up
-// to a sagitta. It ADDS angles and moves none, so the polygon stays inscribed and no existing
-// vertex shifts — a circle nothing touches samples exactly as sampleCircle does.
-func sampleCircleThrough(c *Circle, touches []math.Point2) []math.Point2 {
-	angles := make([]float64, 0, circleSamples+len(touches))
-	for i := range circleSamples {
-		angles = append(angles, 2*stdmath.Pi*float64(i)/float64(circleSamples))
-	}
-	for _, p := range touches {
-		if a, ok := circleTouchAngle(c, p); ok {
-			angles = append(angles, a)
-		}
-	}
-	return circlePointsAt(c, dedupeAngles(angles))
-}
-
-// circleTouchAngle returns p's angle about the circle's centre when p lies on the circle.
-func circleTouchAngle(c *Circle, p math.Point2) (float64, bool) {
-	dx, dy := float64(p.X-c.Center.X), float64(p.Y-c.Center.Y)
-	d := stdmath.Hypot(dx, dy)
-	if stdmath.Abs(d-float64(c.Radius)) > onCircleTol {
-		return 0, false
-	}
-	a := stdmath.Atan2(dy, dx)
-	if a < 0 {
-		a += 2 * stdmath.Pi
-	}
-	return a, true
-}
-
-// dedupeAngles sorts and drops angles closer than the smallest gap worth a vertex, so a touch that
-// lands on (or beside) a uniform sample does not add a degenerate zero-length segment.
-func dedupeAngles(angles []float64) []float64 {
-	sort.Float64s(angles)
-	out := angles[:0]
-	for i, a := range angles {
-		if i == 0 || a-out[len(out)-1] > angleMergeTol {
-			out = append(out, a)
-		}
-	}
-	// the wrap-around pair can also collide once the last angle nears 2*pi
-	if len(out) > 1 && 2*stdmath.Pi-out[len(out)-1]+out[0] <= angleMergeTol {
-		out = out[:len(out)-1]
-	}
-	return out
-}
-
-// angleMergeTol is the angular spacing below which two samples are the same vertex: it is
-// arrMergeTol at unit radius, the same grid the arrangement merges nodes on.
-const angleMergeTol = arrMergeTol // tol:calibrated — angular sample merge; see arrMergeTol
-
-// circlePointsAt evaluates the circle at each angle.
-func circlePointsAt(c *Circle, angles []float64) []math.Point2 {
-	pts := make([]math.Point2, len(angles))
-	for i, a := range angles {
-		pts[i] = math.P2(c.Center.X+c.Radius*math.Scalar(stdmath.Cos(a)), c.Center.Y+c.Radius*math.Scalar(stdmath.Sin(a)))
-	}
-	return pts
-}
-
-// sampleCircleN is sampleCircle at caller-chosen density (region properties
-// scale it with the requested accuracy — M06-F08, #623).
-func sampleCircleN(c *Circle, n int) []math.Point2 {
-	pts := make([]math.Point2, n)
-	for i := range pts {
-		a := 2 * stdmath.Pi * float64(i) / float64(n)
-		pts[i] = math.P2(c.Center.X+c.Radius*stdmath.Cos(a), c.Center.Y+c.Radius*stdmath.Sin(a))
-	}
-	return pts
 }
